requirements.txt
streamlit
pandas
yfinance
openai
plotly
python-dotenv
